overhaul in preperation for deploying virt infra for the ad infra

+++ bin/decrypt.sh
@@ -1,5 +1,7 @@
-ls | grep '\.gpg$' &>/dev/null || exit  
-for file in *.gpg; do
-  gpg --decrypt "$file" > "${file:0:-4}"
+mapfile -t files < <(find . -name "*.secrets.*")
+[ "${#files[@]}" -eq 0 ] && exit
+
+for file in "${files[@]}"; do
+  sops --output-type json --decrypt "$file" > "${file:0:-4}"
   rm "$file"
 done
